update Detail pages
