+++ package.json
@@ -34,7 +34,6 @@
     "react-slick": "^0.29.0",
     "slick-carousel": "^1.8.1",
     "styled-system": "^5.1.5",
-    "theme-ui": "^0.15.4",
     "typescript": "^4.9.5",
     "uninstall": "^0.0.0",
     "web-vitals": "^2.1.3"
