+++ src/pages/ActorDetails.tsx
@@ -1,8 +1,8 @@
 import {
   DetailsDescr,
   DetailsImg,
+  DetailsHomePage,
   DetailsName,
-  Bold,
 } from 'components/DetailsComps/DetailsComps';
 import {
   StyledBtn,
@@ -18,25 +18,25 @@ import {
 } from 'components/Box/Box';
 import { useState, useEffect, useRef } from 'react';
 import { Outlet, useLocation, useParams } from 'react-router-dom';
-import { fetchMovieById } from 'services/api';
+import { fetchActorById } from 'services/api';
 import PageError from 'pages/PageError';
 import Modal from 'components/Modal/Modal';
 import imageplaceholder from 'images/noposter.jpg';
+import { formatDateEn, formatDateUa } from 'services/formatDate';
 import { Suspense } from 'react';
 import { ThreeCircles } from 'react-loader-spinner';
 import { t } from 'i18next';
-import { formatRuntimeEn, formatRuntimeUa } from 'services/formatRuntime';
-import { ImStarHalf } from 'react-icons/im';
+import { ActorDetailsType, IActor } from 'components/types';
 
-export default function MovieDetails({ toggleMovieInLibrary, movies }) {
-  const [movieItem, setMovieItem] = useState(null);
+export default function ActorDetails({ actors, toggleActorsInAlbum }: ActorDetailsType) {
+  const [personInfo, setPersonInfo] = useState<IActor | null>(null);
   const [error, setError] = useState(false);
   const location = useLocation();
-  const params = useParams();
-  const backLink = useRef(location.state?.from ?? '/');
+  const params = useParams() as any;
   const [saved, setSaved] = useState(
-    movies.includes(params.movieId) ? true : false
+    actors.includes(params.actorId) ? true : false
   );
+  const backLink = useRef(location.state?.from ?? '/');
   const [showModal, setShowModal] = useState(false);
   const lang = t('lang');
 
@@ -44,25 +44,25 @@ export default function MovieDetails({ toggleMovieInLibrary, movies }) {
     setShowModal(!showModal);
   };
 
-  const toggleSaveBtn = () => {
-    setSaved(!saved);
-  };
-
-  const handleSaveToLib = () => {
-    toggleMovieInLibrary(params.movieId);
-    toggleSaveBtn();
-  };
-
   useEffect(() => {
-    fetchMovieById(params.movieId, lang)
+    fetchActorById(params.actorId, lang)
       .then(data => {
-        setMovieItem(data);
-        document.title = `My Movies | ${data.title}`;
+        setPersonInfo(data);
+        document.title = `My Movies | ${data.name}`;
       })
       .catch(error => {
         setError(true);
       });
-  }, [lang, params.movieId]);
+  }, [lang, params.actorId]);
+
+  const toggleSaveBtn = () => {
+    setSaved(!saved);
+  };
+
+  const handleSaveToAlbum = () => {
+    toggleActorsInAlbum(params.actorId);
+    toggleSaveBtn();
+  };
 
   return (
     <PageWrap>
@@ -72,65 +72,60 @@ export default function MovieDetails({ toggleMovieInLibrary, movies }) {
 
       {error && <PageError />}
 
-      {movieItem && (
+      {personInfo && (
         <>
-          <DetailsName>{movieItem.title}</DetailsName>
+          <DetailsName>{personInfo.name}</DetailsName>
           <MainInfo>
             <ImgThumb>
               <DetailsImg
                 width="200"
                 src={
-                  movieItem.poster_path
-                    ? `https://image.tmdb.org/t/p/w200/${movieItem.poster_path}`
+                  personInfo.profile_path
+                    ? `https://image.tmdb.org/t/p/w200/${personInfo.profile_path}`
                     : imageplaceholder
                 }
-                alt={`${movieItem.title}`}
+                alt={`${personInfo.name}`}
                 onClick={toggleModal}
               />
             </ImgThumb>
 
             <Box>
-              <DetailsDescr>
-                {(movieItem.release_date ?? movieItem.first_air_date).slice(
-                  0,
-                  4
-                )}
-              </DetailsDescr>
-
-              <DetailsDescr>
-                {movieItem.genres.map(genre => genre.name).join(', ')}
-              </DetailsDescr>
-
-              {movieItem.runtime > 0 && (
+              {personInfo.birthday && (
                 <DetailsDescr>
+                  {t('actor.birth')}
                   {lang === 'uk-UA'
-                    ? `${formatRuntimeUa(movieItem.runtime)}`
-                    : `${formatRuntimeEn(movieItem.runtime)}`}
+                    ? `${formatDateUa(personInfo.birthday)}`
+                    : `${formatDateEn(personInfo.birthday)}`}
                 </DetailsDescr>
               )}
-
-              {movieItem.vote_average > 0 && (
+              {personInfo.deathday && (
                 <DetailsDescr>
-                  <ImStarHalf />{' '}
-                  <Bold bold={movieItem.vote_average > 7}>
-                    {movieItem.vote_average.toFixed(1)}
-                  </Bold>
-                  /10
+                  {t('actor.death')}
+                  {lang === 'uk-UA'
+                    ? `${formatDateUa(personInfo.deathday)}`
+                    : `${formatDateEn(personInfo.deathday)}`}
+                </DetailsDescr>
+              )}
+              {personInfo.homepage && (
+                <DetailsDescr>
+                  {t('actor.homepage')}
+                  <DetailsHomePage
+                    href={personInfo.homepage}
+                    target="_blank"
+                    rel="noreferrer"
+                  >
+                    {personInfo.homepage}
+                  </DetailsHomePage>
                 </DetailsDescr>
               )}
 
               <BtnContainer>
-                <StyledLinkBtn to="overview" state={movieItem.overview}>
-                  {t('buttons.overview')}
+                <StyledLinkBtn to="biography" state={personInfo.biography}>
+                  {t('buttons.biography')}
                 </StyledLinkBtn>
-                <StyledLinkBtn to="cast">{t('buttons.cast')}</StyledLinkBtn>
-                <StyledLinkBtn to="reviews">
-                  {t('buttons.reviews')}
-                </StyledLinkBtn>
-                <StyledLinkBtn to="trailer">
-                  {t('buttons.trailer')}
-                </StyledLinkBtn>
-                <StyledBtn onClick={handleSaveToLib} saved={saved}>
+                <StyledLinkBtn to="movies">{t('buttons.movies')}</StyledLinkBtn>
+                <StyledLinkBtn to="images">{t('buttons.photo')}</StyledLinkBtn>
+                <StyledBtn onClick={handleSaveToAlbum} saved={saved}>
                   {saved ? t('buttons.saved') : t('buttons.save')}
                 </StyledBtn>
               </BtnContainer>
@@ -158,11 +153,11 @@ export default function MovieDetails({ toggleMovieInLibrary, movies }) {
               <img
                 width="100%"
                 src={
-                  movieItem.poster_path
-                    ? `https://image.tmdb.org/t/p/w500/${movieItem.poster_path}`
+                  personInfo.profile_path
+                    ? `https://image.tmdb.org/t/p/w500${personInfo.profile_path}`
                     : imageplaceholder
                 }
-                alt={`${movieItem.original_title}`}
+                alt={personInfo.name}
               />
             </Modal>
           )}
