+++ src/pages/MovieDetails.tsx
@@ -1,8 +1,8 @@
 import {
   DetailsDescr,
   DetailsImg,
-  DetailsHomePage,
   DetailsName,
+  Bold,
 } from 'components/DetailsComps/DetailsComps';
 import {
   StyledBtn,
@@ -18,24 +18,26 @@ import {
 } from 'components/Box/Box';
 import { useState, useEffect, useRef } from 'react';
 import { Outlet, useLocation, useParams } from 'react-router-dom';
-import { fetchActorById } from 'services/api';
+import { fetchMovieById } from 'services/api';
 import PageError from 'pages/PageError';
 import Modal from 'components/Modal/Modal';
 import imageplaceholder from 'images/noposter.jpg';
-import { formatDateEn, formatDateUa } from 'services/formatDate';
 import { Suspense } from 'react';
 import { ThreeCircles } from 'react-loader-spinner';
 import { t } from 'i18next';
+import { formatRuntimeEn, formatRuntimeUa } from 'services/formatRuntime';
+import { ImStarHalf } from 'react-icons/im';
+import { MovieDetailsType, IMovie } from 'components/types';
 
-export default function ActorDetails({ actors, toggleActorsInAlbum }) {
-  const [personInfo, setPersonInfo] = useState(null);
+export default function MovieDetails({ toggleMovieInLibrary, movies }: MovieDetailsType) {
+  const [movieItem, setMovieItem] = useState<IMovie | null>(null);
   const [error, setError] = useState(false);
   const location = useLocation();
-  const params = useParams();
-  const [saved, setSaved] = useState(
-    actors.includes(params.actorId) ? true : false
-  );
+  const params = useParams() as any;
   const backLink = useRef(location.state?.from ?? '/');
+  const [saved, setSaved] = useState(
+    movies.includes(params.movieId) ? true : false
+  );
   const [showModal, setShowModal] = useState(false);
   const lang = t('lang');
 
@@ -43,25 +45,25 @@ export default function ActorDetails({ actors, toggleActorsInAlbum }) {
     setShowModal(!showModal);
   };
 
-  useEffect(() => {
-    fetchActorById(params.actorId, lang)
-      .then(data => {
-        setPersonInfo(data);
-        document.title = `My Movies | ${data.name}`;
-      })
-      .catch(error => {
-        setError(true);
-      });
-  }, [lang, params.actorId]);
-
   const toggleSaveBtn = () => {
     setSaved(!saved);
   };
 
-  const handleSaveToAlbum = () => {
-    toggleActorsInAlbum(params.actorId);
+  const handleSaveToLib = () => {
+    toggleMovieInLibrary(params.movieId);
     toggleSaveBtn();
   };
+
+  useEffect(() => {
+    fetchMovieById(params.movieId, lang)
+      .then(data => {
+        setMovieItem(data);
+        document.title = `My Movies | ${data.title}`;
+      })
+      .catch(error => {
+        setError(true);
+      });
+  }, [lang, params.movieId]);
 
   return (
     <PageWrap>
@@ -71,60 +73,65 @@ export default function ActorDetails({ actors, toggleActorsInAlbum }) {
 
       {error && <PageError />}
 
-      {personInfo && (
+      {movieItem && (
         <>
-          <DetailsName>{personInfo.name}</DetailsName>
+          <DetailsName>{movieItem.title}</DetailsName>
           <MainInfo>
             <ImgThumb>
               <DetailsImg
                 width="200"
                 src={
-                  personInfo.profile_path
-                    ? `https://image.tmdb.org/t/p/w200/${personInfo.profile_path}`
+                  movieItem.poster_path
+                    ? `https://image.tmdb.org/t/p/w200/${movieItem.poster_path}`
                     : imageplaceholder
                 }
-                alt={`${personInfo.name}`}
+                alt={`${movieItem.title}`}
                 onClick={toggleModal}
               />
             </ImgThumb>
 
             <Box>
-              {personInfo.birthday && (
+              <DetailsDescr>
+                {(movieItem.release_date ?? movieItem.first_air_date).slice(
+                  0,
+                  4
+                )}
+              </DetailsDescr>
+
+              <DetailsDescr>
+                {movieItem.genres.map(genre => genre.name).join(', ')}
+              </DetailsDescr>
+
+              {movieItem.runtime > 0 && (
                 <DetailsDescr>
-                  {t('actor.birth')}
                   {lang === 'uk-UA'
-                    ? `${formatDateUa(personInfo.birthday)}`
-                    : `${formatDateEn(personInfo.birthday)}`}
+                    ? `${formatRuntimeUa(movieItem.runtime)}`
+                    : `${formatRuntimeEn(movieItem.runtime)}`}
                 </DetailsDescr>
               )}
-              {personInfo.deathday && (
+
+              {movieItem.vote_average > 0 && (
                 <DetailsDescr>
-                  {t('actor.death')}
-                  {lang === 'uk-UA'
-                    ? `${formatDateUa(personInfo.deathday)}`
-                    : `${formatDateEn(personInfo.deathday)}`}
-                </DetailsDescr>
-              )}
-              {personInfo.homepage && (
-                <DetailsDescr>
-                  {t('actor.homepage')}
-                  <DetailsHomePage
-                    href={personInfo.homepage}
-                    target="_blank"
-                    rel="noreferrer"
-                  >
-                    {personInfo.homepage}
-                  </DetailsHomePage>
+                  <ImStarHalf />{' '}
+                  <Bold bold={movieItem.vote_average > 7}>
+                    {movieItem.vote_average.toFixed(1)}
+                  </Bold>
+                  /10
                 </DetailsDescr>
               )}
 
               <BtnContainer>
-                <StyledLinkBtn to="biography" state={personInfo.biography}>
-                  {t('buttons.biography')}
+                <StyledLinkBtn to="overview" state={movieItem.overview}>
+                  {t('buttons.overview')}
                 </StyledLinkBtn>
-                <StyledLinkBtn to="movies">{t('buttons.movies')}</StyledLinkBtn>
-                <StyledLinkBtn to="images">{t('buttons.photo')}</StyledLinkBtn>
-                <StyledBtn onClick={handleSaveToAlbum} saved={saved}>
+                <StyledLinkBtn to="cast">{t('buttons.cast')}</StyledLinkBtn>
+                <StyledLinkBtn to="reviews">
+                  {t('buttons.reviews')}
+                </StyledLinkBtn>
+                <StyledLinkBtn to="trailer">
+                  {t('buttons.trailer')}
+                </StyledLinkBtn>
+                <StyledBtn onClick={handleSaveToLib} saved={saved}>
                   {saved ? t('buttons.saved') : t('buttons.save')}
                 </StyledBtn>
               </BtnContainer>
@@ -152,11 +159,11 @@ export default function ActorDetails({ actors, toggleActorsInAlbum }) {
               <img
                 width="100%"
                 src={
-                  personInfo.profile_path
-                    ? `https://image.tmdb.org/t/p/w500${personInfo.profile_path}`
+                  movieItem.poster_path
+                    ? `https://image.tmdb.org/t/p/w500/${movieItem.poster_path}`
                     : imageplaceholder
                 }
-                alt={personInfo.name}
+                alt={`${movieItem.original_title}`}
               />
             </Modal>
           )}
